blog/backend/requirements.txt
asgiref==3.8.1
dj-rest-auth==6.0.0
Django==5.0.4
django-allauth==0.62.1
djangorestframework==3.15.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
